requirements.txt
appnope==0.1.4
asttokens==2.4.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.1
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pyzmq==26.2.0
setuptools==74.1.2
six==1.16.0
stack-data==0.6.3
sympy==1.13.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
